// src/app/models/People.ts
export class People {
    name: string;
    height: string;
    mass: string;
    // tslint:disable-next-line:variable-name
    hair_color: string;
    // tslint:disable-next-line:variable-name
    skin_color: string;
    // tslint:disable-next-line:variable-name
    eye_color: string;
    // tslint:disable-next-line:variable-name
    birth_year: string;
    gender: string;
    homeworld: string;
    films: string[];
    species: string[];
    vehicles: string[];
    starships: string[];
    created: string;
    edited: string;
    url: string;
}